delete button add to table
